Moved command driver to cnab-go
