+++ driver/command/command_nix.go
@@ -1,15 +1,16 @@
-// +build windows
+// +build !windows
 
-package driver
+package command
 
 import (
+	"fmt"
 	"os"
 	"os/exec"
 )
 
 // CheckDriverExists checks to see if the named driver exists
-func (d *CommandDriver) CheckDriverExists() bool {
-	cmd := exec.Command("where", d.cliName())
+func (d *Driver) CheckDriverExists() bool {
+	cmd := exec.Command("/bin/sh", "-c", fmt.Sprintf("command -v %s", d.cliName()))
 	cmd.Env = os.Environ()
 	if err := cmd.Run(); err != nil {
 		return false
